Add GVRSceneRenderTexture instead of proprietary APIs in GVRContext
- still requires creation on the GL thread; refactoring GVRRenderTexture
will be done separately, if necessary

GearVRf-DCO-1.0-Signed-off-by: Mihail Marinov <m.marinov@samsung.com>
